Block users from deleting other users' messages. Modified messages_destroy(message_id) in app.py.

+++ app.py
@@ -346,8 +346,13 @@ def messages_destroy(message_id):
         return redirect("/")
 
     msg = Message.query.get(message_id)
-    db.session.delete(msg)
-    db.session.commit()
+    # verify user is trying to delete own message
+    if g.user == msg.user:
+        db.session.delete(msg)
+        db.session.commit()
+    else:
+        flash("Access unauthorized.", "danger")
+        return redirect("/")
 
     return redirect(f"/users/{g.user.id}")
